Move old code to old dirs
