+++ .env.example
@@ -1,5 +1,6 @@
 DISCORD_BOT_TOKEN=xxxxx
 DISCORD_SERVER_ID=xxxxxxx
+DISCORD_CLIENT_ID=xxxxxx
 DISCORD_COMMAND_PREFIX=!
 DATABASE_URL=xxxxx
 DISCORD_ADMIN_ROLE_ID=xxxxxx
