+++ core/logging.js
@@ -1,3 +1,6 @@
-const logger = require('pino')();
+const pino = require('pino');
+
+// Configure logger
+const logger = pino();
 
 module.exports = logger;
