+++ log.json
@@ -1,0 +1,1 @@
+{"level":30,"time":1653674514413,"pid":25079,"hostname":"frank-pc","msg":"Ready!"}
