+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "itws-discord-bot",
   "version": "2.0.0",
-  "description": "Test Discord Role Bot",
+  "description": "Discord bot that manages the RPI ITWS Discord server.",
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
@@ -10,6 +10,7 @@
   "author": "Simar Singh",
   "license": "ISC",
   "dependencies": {
+    "@discordjs/builders": "^0.13.0",
     "@discordjs/rest": "^0.4.1",
     "discord-api-types": "^0.33.0",
     "discord.js": "^13.7.0",
